added new components as part of refactoring core components
