+++ likeminds-feed-reactnative-integration/store/reducers/feedReducer.ts
@@ -34,6 +34,7 @@ import {
   SET_REPORT_MODEL_STATUS_IN_POST_DETAIL,
   SET_FLOW_TO_CREATE_POST_SCREEN,
   SET_FLOW_TO_CAROUSEL_SCREEN,
+  MAPPED_TOPICS_FROM_UNIVERSAL_FEED_SCREEN,
 } from "../types/types";
 import { LMPostUI } from "../../models";
 import Styles from "../../constants/Styles";
@@ -44,6 +45,7 @@ export interface FeedReducerState {
   reportTags: {};
   autoPlayVideoPostId: "";
   topics: {};
+  mappedTopics: {};
   selectedTopicsForUniversalFeedScreen: [];
   selectedTopicsForCreatePostScreen: [];
   selectedTopicsFromUniversalFeedScreen: [];
@@ -62,6 +64,7 @@ export const initialState: FeedReducerState = {
   reportTags: {},
   autoPlayVideoPostId: "",
   topics: {},
+  mappedTopics: {},
   selectedTopicsForUniversalFeedScreen: [],
   selectedTopicsForCreatePostScreen: [],
   selectedTopicsFromUniversalFeedScreen: [],
@@ -140,6 +143,13 @@ export const feedReducer = (state = initialState, action) => {
     case SET_TOPICS: {
       const { topics = {} } = action.body;
       return { ...state, topics: topics };
+    }
+    case MAPPED_TOPICS_FROM_UNIVERSAL_FEED_SCREEN: {
+      const { topics = {} } = action.body;
+      return {
+        ...state,
+        mappedTopics: topics,
+      };
     }
     case UNIVERSAL_FEED_SUCCESS: {
       const { users = {} } = action.body;
