+++ likeminds-feed-reactnative-integration/store/types/types.ts
@@ -31,6 +31,8 @@ export const CLEAR_SELECTED_TOPICS_FOR_CREATE_POST_SCREEN =
   "CLEAR_SELECTED_TOPICS_FOR_CREATE_POST_SCREEN";
 export const SELECTED_TOPICS_FROM_UNIVERSAL_FEED_SCREEN =
   "SELECTED_TOPICS_FROM_UNIVERSAL_FEED_SCREEN";
+export const MAPPED_TOPICS_FROM_UNIVERSAL_FEED_SCREEN =
+  "MAPPED_TOPICS_FROM_UNIVERSAL_FEED_SCREEN";
 export const CLEAR_SELECTED_TOPICS_FROM_UNIVERSAL_FEED_SCREEN =
   "CLEAR_SELECTED_TOPICS_FROM_UNIVERSAL_FEED_SCREEN";
